backend/requirements.txt
# Core Dependencies
fastapi
uvicorn
pydantic
gql[aiohttp]
web3
requests
python-dotenv
